# Small micro-optimizations around set performance