--- a/internal/services/v1/debug_test.go
+++ b/internal/services/v1/debug_test.go
@@ -49,7 +49,7 @@ func expectDebugFrames(permissionNames ...string) rda {
 		for _, sp := range debugInfo.Check.GetSubProblems().Traces {
 			for _, permissionName := range permissionNames {
 				if sp.Permission == permissionName {
-					found.Add(permissionName)
+					found.Insert(permissionName)
 				}
 			}
 		}