--- a/pkg/tuple/onrset.go
+++ b/pkg/tuple/onrset.go
@@ -3,8 +3,9 @@ package tuple
 import (
 	"maps"
 
-	core "github.com/authzed/spicedb/pkg/proto/core/v1"
 	expmaps "golang.org/x/exp/maps"
+
+	core "github.com/authzed/spicedb/pkg/proto/core/v1"
 )
 
 // ONRSet is a set of ObjectAndRelation's.