Before the changes:

```
BenchmarkAdd-10          	16160137	       128.5 ns/op
BenchmarkCopy-10         	   10000	    465419 ns/op
BenchmarkHas-10          	47101461	        49.75 ns/op
BenchmarkDelete-10       	27999956	        82.98 ns/op
BenchmarkIntersect-10    	   10000	    772631 ns/op
BenchmarkSubtract-10     	   10000	    933186 ns/op
```

After the changes:

```
BenchmarkAdd-10          	16139670	       130.6 ns/op
BenchmarkInsert-10       	18618121	       112.0 ns/op
BenchmarkCopy-10         	   10000	    108095 ns/op
BenchmarkHas-10          	46172048	        51.45 ns/op
BenchmarkDelete-10       	31368414	        60.33 ns/op
BenchmarkIntersect-10    	   10000	    386777 ns/op
BenchmarkSubtract-10     	   10000	    474429 ns/op
BenchmarkAsSlice-10      	   14143	    136824 ns/op
BenchmarkEqual-10        	   10000	    275370 ns/op
```

The big changes are in `Copy`, `Subtract` and `Equal`, all of which are now using the native runtime-provided `clone` for maps.

The change from `Add` to `Insert` (where applicable) and from `Remove` to `Delete` also saves 10-20ns per call